Fix SignedIdentifer XML wrapper name (#4709)
* Fix SignedIdentifer XML wrapper name

* Simply client builders

+++ sdk/storage/azure-storage-file/src/main/java/com/azure/storage/file/ShareAsyncClient.java
@@ -528,7 +528,7 @@ public class ShareAsyncClient {
     }
 
     private Response<ShareStatistics> mapGetStatisticsResponse(SharesGetStatisticsResponse response) {
-        ShareStatistics shareStatistics = new ShareStatistics(response.value().shareUsageBytes() / 1024);
+        ShareStatistics shareStatistics = new ShareStatistics((int) (response.value().shareUsageBytes() / 1024));
 
         return new SimpleResponse<>(response.request(), response.statusCode(), response.headers(), shareStatistics);
     }
